refactor(vo): DTO -> VO로 패키지명 수정
- 데이터 전달을 위해 만들어진 클래스들이 아니기 때문에 수정
- InsertPrice 삭제 후 Price로 통일
  유저가 int 범위를 넘어설 만큼 돈이 많다고 가정

Related:
- https://github.com/next-step/java-lotto-kakao/pull/24#discussion_r554639809
- https://github.com/next-step/java-lotto-kakao/pull/24#discussion_r554458484

+++ src/main/java/lotto/view/OutputView.java
@@ -10,7 +10,7 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 public class OutputView {
-    public void printNumberOfLottoTickets(int ticketCount, int change) {
+    public void printNumberOfLottoTickets(long ticketCount, long change) {
         System.out.printf("%d개를 구매했습니다.\n", ticketCount);
 
         if (hasChange(change)) {
@@ -50,7 +50,7 @@ public class OutputView {
         System.out.println(message);
     }
 
-    private boolean hasChange(int change) {
+    private boolean hasChange(long change) {
         return change > 0;
     }
 }
